worker/jobs/index: Extract `SquashIndex` into dedicated module
